feat: complete Enterprise GEO/SEO framework injection and accessibility optimization
