+++ app/insights/cognitive-architecture-blueprint/page.tsx
@@ -1,37 +1,13 @@
+"use client"
+
 import { TechnicalCitations } from "@/components/insights/technical-citations"
 import { StatisticalSignal } from "@/components/insights/statistical-signal"
 import { ExecutiveSummary } from "@/components/insights/executive-summary"
 import { FAQSection } from "@/components/insights/faq-section"
 import { AuthorBio } from "@/components/insights/author-bio"
 import { BreadcrumbSchema } from "@/components/insights/breadcrumb-schema"
-import type { Metadata } from 'next'
 
-export const metadata: Metadata = {
-  title: 'The Cognitive Architecture Blueprint | ADI Methodology',
-  description: 'How Inner G Complete applies the PMI-certified CPMAI framework to architect governance-first Aesthetic Domain Intelligence (ADI) models.',
-  keywords: ['CPMAI framework', 'cognitive architecture', 'ADI methodology', 'AI project management'],
-  openGraph: {
-    title: 'The Cognitive Architecture Blueprint | Inner G Complete',
-    type: 'article',
-    url: 'https://innergcomplete.com/insights/cognitive-architecture-blueprint',
-    siteName: 'Inner G Complete Agency',
-    images: [
-      {
-        url: '/cpmai_framework_cover.png',
-        width: 1200,
-        height: 630,
-      },
-    ],
-  },
-  twitter: {
-    card: "summary_large_image",
-    title: "Cognitive Architecture Blueprint",
-    images: ['/cpmai_framework_cover.png'],
-  },
-  alternates: {
-    canonical: "https://innergcomplete.com/insights/cognitive-architecture-blueprint",
-  },
-}
+
 
 
 import { Navbar } from "@/components/layout/navbar"
@@ -228,6 +204,7 @@ export default function CognitiveArchitectureBlueprint() {
           })
         }}
       />
+      <BreadcrumbSchema slug="cognitive-architecture-blueprint" title="Cognitive Architecture Blueprint | Technical View | Inner G Complete" />
       <Navbar />
 
       <article className="relative flex-1">
@@ -296,10 +273,10 @@ export default function CognitiveArchitectureBlueprint() {
                 </div>
               </div>
               <div className="ml-auto flex gap-3">
-                <Button variant="outline" size="icon" className="rounded-full h-10 w-10 border-border">
+                <Button variant="outline" size="icon" className="rounded-full h-10 w-10 border-border" aria-label="Share article">
                   <Share2 className="h-4 w-4" />
                 </Button>
-                <Button variant="outline" size="icon" className="rounded-full h-10 w-10 border-border">
+                <Button variant="outline" size="icon" className="rounded-full h-10 w-10 border-border" aria-label="Print article">
                   <Printer className="h-4 w-4" />
                 </Button>
               </div>
