+++ app/insights/mindbody-sovereign-intelligence-audit/page.tsx
@@ -75,6 +75,7 @@ export default function MindbodySovereignIntelligenceAudit() {
           })
         }}
       />
+      <BreadcrumbSchema slug="mindbody-sovereign-intelligence-audit" title="MindBody's Intelligence Ceiling | Strategic View | Inner G Complete" />
       <Navbar />
 
       <article className="relative flex-1">
@@ -143,10 +144,10 @@ export default function MindbodySovereignIntelligenceAudit() {
                 </div>
               </div>
               <div className="ml-auto flex gap-3">
-                <Button variant="outline" size="icon" className="rounded-full h-10 w-10 border-border">
+                <Button variant="outline" size="icon" className="rounded-full h-10 w-10 border-border" aria-label="Share article">
                   <Share2 className="h-4 w-4" />
                 </Button>
-                <Button variant="outline" size="icon" className="rounded-full h-10 w-10 border-border">
+                <Button variant="outline" size="icon" className="rounded-full h-10 w-10 border-border" aria-label="Print article">
                   <Printer className="h-4 w-4" />
                 </Button>
               </div>
