+++ app/insights/the-sovereign-intelligence-layer/page.tsx
@@ -1,37 +1,13 @@
+"use client"
+
 import { TechnicalCitations } from "@/components/insights/technical-citations"
 import { StatisticalSignal } from "@/components/insights/statistical-signal"
 import { ExecutiveSummary } from "@/components/insights/executive-summary"
 import { FAQSection } from "@/components/insights/faq-section"
 import { AuthorBio } from "@/components/insights/author-bio"
 import { BreadcrumbSchema } from "@/components/insights/breadcrumb-schema"
-import type { Metadata } from 'next'
 
-export const metadata: Metadata = {
-  title: 'The Sovereign Intelligence Layer | Core Vision',
-  description: 'Why the enterprise that builds a proprietary Artificial Domain Intelligence creates an unassailable competitive moat.',
-  keywords: ['sovereign intelligence layer', 'Artificial Domain Intelligence vision', 'enterprise grooming AI moat'],
-  openGraph: {
-    title: 'The Sovereign Intelligence Layer: Why ADI Wins',
-    type: 'article',
-    url: 'https://innergcomplete.com/insights/the-sovereign-intelligence-layer',
-    siteName: 'Inner G Complete Agency',
-    images: [
-      {
-        url: '/adi_sovereign_layer_cover_1776108008232.png',
-        width: 1200,
-        height: 630,
-      },
-    ],
-  },
-  twitter: {
-    card: "summary_large_image",
-    title: "The Sovereign Intelligence Layer",
-    images: ['/adi_sovereign_layer_cover_1776108008232.png'],
-  },
-  alternates: {
-    canonical: "https://innergcomplete.com/insights/the-sovereign-intelligence-layer",
-  },
-}
+
 
 
 import { Navbar } from "@/components/layout/navbar"
@@ -121,6 +97,7 @@ export default function SovereignIntelligenceLayer() {
           })
         }}
       />
+      <BreadcrumbSchema slug="the-sovereign-intelligence-layer" title="The Sovereign Intelligence Layer | Core Vision" />
       <Navbar />
 
       <article className="relative flex-1">
